refactor(folio-creation): remove unused state and parameters
