+++ src/features/folio/context/FolioCreationContext.tsx
@@ -1,8 +1,9 @@
-import React, { createContext, useContext, useState } from 'react';
+import React, { createContext, useContext, useState, useCallback } from 'react';
 
 interface FolioCreationContextType {
   name: string;
   setName: (name: string) => void;
+  reset: () => void;
 }
 
 const FolioCreationContext = createContext<FolioCreationContextType | undefined>(undefined);
@@ -10,8 +11,10 @@ const FolioCreationContext = createContext<FolioCreationContextType | undefined>
 export const FolioCreationProvider = ({ children }: { children: React.ReactNode }) => {
   const [name, setName] = useState('Untitled');
 
+  const reset = useCallback(() => setName('Untitled'), []);
+
   return (
-    <FolioCreationContext.Provider value={{ name, setName }}>
+    <FolioCreationContext.Provider value={{ name, setName, reset }}>
       {children}
     </FolioCreationContext.Provider>
   );
